# docs/OpenAPI.yml
openapi: 3.0.0
...
servers:
# Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AnorZaken7/LoneStarLudo/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AnorZaken7/LudoMoves/1.0.0
info:
  description: API for the Ludo Lone Star game server.
  version: "2.0.0"
  title: Ludo Lone Star API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /ludo/user:
    get:
      operationId: ludoListUsers
      tags:
        - users
      summary: Get a list of users, optionally filtered.
      parameters:
        - name: userName
          in: query
          description: UserName of a user to search for.
          schema:
            type: string
      responses:
        '200':
          description: A list of user ids matching the request.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '404':
          description: No matching user(s) exist. (For searches only.)
    
    post:
      operationId: ludoCreateUser
      tags:
        - users
      summary: Creates a new user.
      parameters:
        - name: UserName
          in: header
          description: The UserName of the user to create (need to be unique).
          schema:
            type: string
      responses:
        '201':
          description: User created (see Location header).
          #TODO: add auth token?
        '400':
          description: Missing or empty UserName parameter.
        '409':
          description: UserName was not unique.
        '422':
          description: Non-acceptable UserName parameter (see body for details).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          #TODO: add detailed response, e.g. too short/long, profane, etc.
  /ludo/user/{userId}:
    parameters:
      - $ref: '#/components/parameters/userIdPath'
    
    get:
      operationId: ludoGetUser
      tags: 
        - users
      summary: Get information about a user.
      responses:
        '200':
          description: A UserInfo object with information about the user.
          content:
            application/json:
              items:
                $ref: '#/components/schemas/UserInfo'
        '404':
          description: No such user exists.
      #TODO: add more info / statistics here? or add sub-paths with more info?
    
    #put: ...
    #TODO: add ability to make user changes - requires some form of auth!
    
    #delete: ...
    #TODO: should be a delete operation - requires some form of auth!
  /ludo/lobby:
    get:
      operationId: ludoListLobbies
      tags:
        - users
      summary: Lists Ludo lobbies.
      description: TODO
      parameters:
        - name: show
          in: query
          description: |
            Filters lobbies based on how full they are;
            open - Show open only.
            full - Show full only.
            all - Show both open and full.
            penultimate - Show only lobbies with exactly one unoccupied slot.
          schema:
            type: string
            default: open
            enum: [open, full, all, penultimate]
        - name: userId
          in: query
          description: >
            Filter to lobbies where at least one specified user is present
            (e.g. your friends).
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: A JSON array of LobbyListEntry objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LobbyListEntry'
        '400':
          description: bad input parameter
        '404':
          description: One or more of the specified players does not exists.

    post:
      operationId: ludoCreateLobby
      tags:
        - users
      summary: Creates a new lobby.
      description: Creates a new lobby and joins it as player 0.
      requestBody:
        description: Information required to create a new lobby.
        required: true
        content:
          application/JSON:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CreateLobby'
                - $ref: '#/components/schemas/LoadLobby'
      responses:
        '201': # Created
          description: >
            Returns the id of the newly created game in the Location header.
            The creator of a lobby is always assigned to slot 0 when creating a new game.
        '400':
          description: Player not specified.
        '404':
          description: No such user exists (userId included).
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: SlotCount is out of range or invalid save data.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        #TODO: some form of auth?
  /ludo/lobby/{gameId}:
    parameters:
      - $ref: '#/components/parameters/gameIdPath'
    
    get:
      operationId: ludoGetLobby
      tags: 
        - users
      summary: Get lobby information.
      description: Which players are in the game, can the game still be joined, etc.
      responses:
        '200':
          description: A LobbyInfo object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LobbyInfo'
        '404':
          description: No such game exists.
    
    patch:
      operationId: ludoJoinLobby
      tags:
        - users
      summary: Join a game lobby.
      description: Adds the specified player to a game lobby, returning an auto assigned slot index.
      parameters: 
        - name: UserId
          in: header
          description: The user who wants to join.
          schema:
            type: string
          required: true
      responses:
        '200':
          description: >
            Game joined. Here's your slot index for this game,
            or -1 if not automatically assigned to any slot.
            (Also sent if the player has already joined this game.)
          content:
            text/plain:
              schema:
                type: integer
                format: int32
                minimum: -1
                maximum: 3
          #TODO: return game auth token
        '400':
          description: Player not specified.
        '404':
          description: No such game and / or player exists.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        '409': # "This response is sent when a request conflicts with the current state of the server."
          description: Unable to join game
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
          #TODO: supply exact reason in body (Error object)
    
    post:
      operationId: ludoStartGame
      tags: 
        - players
      summary: Start game.
      description: |
        Starts the game.
        Requires lobby state.
        Requires at least two players.
        Requires that all players are ready.
      responses:
        '201':
          description: Game started (see Location header).
        '404':
          description: No such game exists.
        '409':
          description: Unable to start game.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
      #TODO: require game auth token + 401 & 403 response
    
    delete:
      operationId: ludoLeaveLobby
      tags:
        - players
      summary: Leave lobby.
      description: When the last player leaves, the lobby itself is deleted.
      parameters: 
        - name: UserId
          in: header
          description: The player who wants to leave.
          schema:
            type: string
          required: true
      responses:
        '204':
          description: >
            User has either been removed from the lobby (change),
            or user wasn't a member of this lobby (no change).
        '400':
          description: UserId not specified.
        '404':
          description: No such game and / or user exists.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Game is not in the setup phase.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
      #TODO: require game auth token + 401 & 403 response
  /ludo/lobby/{gameId}/{slot}:
    parameters:
      - $ref: '#/components/parameters/gameIdPath'
      - $ref: '#/components/parameters/slotPath'
    
    get:
      operationId: ludoGetPlayerReady
      tags:
        - users
      summary: Get player and ready state for the specified lobby slot.
      responses:
        '200':
          description: OK
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/PlayerReady'
        '400':
          description: Bad slot index.
        '404':
          description: No such game or slot exists.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Game not in setup stage (i.e. already in progress / ended).
      #TODO: require game auth token + 401 & 403 response

    post:
      operationId: ludoSetSlotPlayer
      tags:
        - players
      summary: Move a user to this slot.
      description: >
        User must already be in the lobby and the slot needs to be empty
        or occupied by someone whom the user has permission to evict.
        (A reservation grants permission to evict those without a reservation.)
        Use slot -1 to evict yourself.
      parameters:
        - name: UserId
          in: header
          description: User to move to this slot.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: OK - Also sent if no change occurred.
        '400':
          description: Bad input parameter.
        '404':
          description: No such game / user / slot exists.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Unable to claim slot.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
      #TODO: require game auth token + 401 & 403 response
    
    put:
      operationId: ludoSetSlotReady
      tags:
        - players
      summary: Sets the ready status of a slot/player.
      description: All players in a lobby need to ready up before a game can start.
      requestBody:
        description: The user occupying this slot to and their new ready status.
        required: true
        content:
          application/JSON:
            schema:
              $ref: '#/components/schemas/PlayerReady'
      responses:
        '204':
          description: OK - Also sent if no change occurred.
        '400':
          description: Bad input parameter.
        '404':
          description: No such game / user / slot exists.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >
            Unable to set ready status
            (e.g. the user found in this slot was not the user specified).
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
      #TODO: require game auth token + 401 & 403 response
...
components:
  parameters:
...
    gameIdPath:
      name: gameId
      in: path
      required: true
      description: Id of a game.
      schema:
        type: string
    slotPath:
      name: slot
      in: path
      required: true
      description: Game specific slot.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 3
...
  schemas:
    Error:
      type: object
      parameters:
        code:
          description: An error code.
          type: integer
          format: int32
        desc:
          description: Optional human readable description.
          type: string
    GameState:
      type: string
      enum: [setup, ingame, finished]
      description: The state of a game / lobby.
    LobbyAccess:
      type: string
      enum: [public, unlisted, friendsOnly, inviteOnly, reservations]
      description: |
        Setting that determines the rules for joining a lobby.
        public - game is publicly listed and anyone can join.
        unlisted - anyone with the gameId can join, but game is not publicly listed.
        friendsOnly - game is publicly listed, but only friends can join. (wip)
        inviteOnly - game is not publicly listed and can only be joined by invite. (wip)
        reservations - public game where anyone can join; however slotes are pre-associated
        with playerIds such that if those players join they are automatically granted their pre-associated
        index, vacating any previous occupant from that slot. (wip)
    CreateLobby:
      type: object
      description: Information required to create a new lobby.
      required:
        - userId
      properties:
        userId:
          description: The user creating the lobby.
          type: string
        slots:
          description: Number of slots.
          type: integer
          format: int32
          minimum: 2
          maximum: 4
          default: 4
        access:
          $ref: '#/components/schemas/LobbyAccess'
        reservations:
          description: An array of all slot reservations in this lobby.
          type: array
          items:
            $ref: '#/components/schemas/LobbyReservation'
    LoadLobby:
      type: object
      description: Information required to create a lobby from a saved game.
      required:
        - save
      properties:
        save:
          $ref:'#/components/schemas/GameSave'
        access:
          $ref: '#/components/schemas/LobbyAccess'
        strict:
          type: boolean
          description: >
            Reservations are always generated for the old participants when loading a save.
            This determines if those reservations are strict or relaxed.
          default: false
    UserInfo:
      type: object
      properties:
        userName:
          type: string
        #TODO: add more info; dateJoined etc...
    PlayerReady:
      type: object
      required:
        - userId
        - ready
      properties:
        userId:
          type: string
        ready:
          type: boolean
    LobbyReservation:
      type: object
      required:
        - player
        - slot
      properties:
        player:
          type: string
        slot:
          description: Slot index, or -1 if the reservation is slot agnostic.
          type: integer
          format: int32
          minimum: -1
          maximum: 3
        strict:
          type: boolean
          description: >
            If the reservation is strict, then no other player may take this slot.
            Conversely a relaxed reservation behaves like a normal slot, except
            that a player that has a reservation to may evict non-reserved players
            to claim it.
          default: false
    LobbyListEntry:
      type: object
      properties:
        gameId:
          type: string
        access:
          $ref: '#/components/schemas/LobbyAccess'
        slots:
          description: >
            An array of players in slot index order.
            Empty slots are represented by empty array items.
          type: array
          items:
            type: string
        others:
          description: >
            An array of players that are not currently assigned to any slot.
          type: array
          items:
            type: string
        #TODO: add reservation info <-------------------------
    LobbyInfo:
      type: object
      properties:
        state:
          $ref: '#/components/schemas/GameState'
        access:
          $ref: '#/components/schemas/LobbyAccess'
        slots:
          description: >
            An array of players and ready status in slot index order.
            Empty slots are represented by empty array items.
          type: array
          items:
            $ref: '#/components/schemas/PlayerReady'
        others:
          description: >
            An array of players that are not currently assigned to any slot.
          type: array
          items:
            type: string
        reservations:
          description: An array of all slot reservations in this lobby.
          type: array
          items:
            $ref: '#/components/schemas/LobbyReservation'
        #TODO: add more properties..?
...
    PieceDistances:
      type: array
      description: |
        An array of piece distances (for a player's pieces).
        These player relative board positions describe how far a piece has moved.
        To translate them into absolute board positions, the length of the board and the index of the player must also be known.
        Assuming the distance is less than the board length (i.e. if the piece is still on the shared competative track) the formula looks like so;
        position = (distance + slot * boardLength / 4) % boardLength
      items:
        type: integer
        format: int32
        minimum: 0
        maximum: 90 # based on maximum allowed boardLength
    BoardState:
      type: array
      description: An array of piece distance arrays (one for each player, i.e. slot index).
      items:
        $ref: '#/components/schemas/PieceDistances'
    TurnSlotDie:
      type: object
      required:
        - turn
        - player
        - die
      properties:
        turn:
          type: integer
          description: Which turn the game is on.
          format: int32
          minimum: 0
        slot:
          type: integer
          description: Slot index of current player.
          format: int32
          minimum: 0
          maximum: 3
        die:
          type: integer
          description: Die roll for the current player / turn.
          format: int32
          minimum: 1
          maximum: 6
    PlayerSlot:
      type: object
      properties:
        player:
          type: string
        slot:
          description: Slot index, or -1 if not assigned to any slot.
          type: integer
          format: int32
          minimum: -1
          maximum: 3
    LobbySave:
      type: object
      required:
        - players
      properties:
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerSlot'
    GameSave:
      type: object
      required:
        - board
        - current
        - lobby
      properties:
        board:
          $ref: '#/components/schemas/BoardState'
        current:
          $ref: '#/components/schemas/TurnSlotDie'
        lobby:
          $ref: '#/components/schemas/LobbySave'
        #TODO is any data missing to make this a complete game save?
...